backend/requirements.txt

django==4.0
psycopg2>=2.8
djangorestframework==3.13.1
django-pghistory==1.5.0

# sqlparse==0.4.2
# zope.event==4.5.0
# zope.interface==5.4.0
# gevent==21.12.0
# greenlet==1.1.2
# asgiref==3.5.2
# backports.zoneinfo==0.2.1
# gunicorn==20.1.0